refactor: change everything to screens and move components into separate folders
BREAKING CHANGES
