+++ components/AutoFill/index.js
@@ -3,7 +3,7 @@ import {
   StyleSheet, Text, TouchableOpacity, View
 } from 'react-native';
 import Autocomplete from 'react-native-autocomplete-input';
-import retrievePuenteAutofillData from '../services/aws';
+import retrievePuenteAutofillData from '../../services/aws';
 
 // Demo base API to get the data for the Autocomplete suggestion
 export default class AutoFill extends Component {
@@ -76,8 +76,8 @@ export default class AutoFill extends Component {
           {fields.length > 0 ? (
             <Text style={styles.infoText}>{query}</Text>
           ) : (
-            <Text style={styles.infoText}>{placeholder}</Text>
-          )}
+              <Text style={styles.infoText}>{placeholder}</Text>
+            )}
         </View>
       </View>
     );
